feat: extract shared types, constants, and utility modules from AppView.tsx
- Created src/lib/types.ts: SessionStatus, ArtifactManifest, DashboardState,
  SessionSummary, LogEntry, FileChange, ArtifactEntry, ManifestPayload,
  ProviderHealth, IterationUsage (96 LOC)
- Created src/lib/ansi.ts: AnsiStyle, PALETTE_256, stripAnsi, rgbStr,
  parseAnsiSegments, renderAnsiToHtml (116 LOC)
- Created src/lib/format.ts: formatTime, formatTimeShort, formatSecs,
  formatDuration, formatDateKey, relativeTime, formatTokenCount, slugify (56 LOC)
- Created src/lib/session.ts: IMAGE_EXT, isRecord, str, numStr, toSession,
  isImageArtifact, artifactUrl, deriveProviderHealth (104 LOC)
- Created src/lib/parse.ts: SIGNIFICANT_EVENTS, parseLogLine,
  extractModelFromOutput, parseDurationSeconds, computeAvgDuration,
  parseManifest, findBaselineIterations, extractIterationUsage (145 LOC)
- Updated AppView.tsx: replaced all extracted definitions with imports,
  added re-exports for all previously-exported symbols

All 129 tests pass, npm run type-check clean, all new files < 150 LOC.

Aloop-Agent: build
Aloop-Iteration: 2
Aloop-Session: orchestrator-20260321-172932-issue-151-20260323-134315
